Added an example ZMQ server written in Fortran.
Added a ZeroMQ server written in Fortran.  This example binds to
a fixed port (5555) on the local host and expects requests to be
composed of two frames.  The first frame contains a string with
the name of a method to invoke.  The second frame contains the
arguments encoded into a JSON string.  Two methods are provided.
An included Python script demonstrates use.

The program depends on the C ZMQ library and on Fortran-JSON.  An
included Gradle build script simply executes `make`.  The makefile
hard codes the header and library paths to these dependencies.